Initial webpack structure
