+++ .gitignore
@@ -1,2 +1,3 @@
 node_modules
 *.zip
+chrome-extension/js/*.js
