+++ chrome-extension/background.html
@@ -1,12 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-  <script src="config.json"></script>
-  <script src="/js/settings.js"></script>
-  <script src="/js/xhr.js"></script>
-  <script src="/js/lastfm/api.md5.js"></script>
-  <script src="/js/lastfm/api.js"></script>
-  <script src="/js/lastfm.obj.js"></script>
+  <script type="text/javascript" src="/js/commons.js"></script>
   <script src="/js/background.js"></script>
 </head>
 <body>
