--- chrome-extension/js/lastfm.obj.js
@@ -1,4 +1,0 @@
-var lastfm = new LastFM({
-  apiKey    : CONFIG.lastFM.key,
-  apiSecret : CONFIG.lastFM.sig
-});
