+++ chrome-extension/options.html
@@ -5,12 +5,7 @@
   <link href="/css/generic.css" rel="stylesheet" type="text/css" />
   <link href="/css/options.css" rel="stylesheet" type="text/css" />
   <title>UltraFM - options..</title>
-  <script src="/config.json"></script>
-  <script src="/js/settings.js"></script>
-  <script src="/js/xhr.js"></script>
-  <script src="/js/lastfm/api.md5.js"></script>
-  <script src="/js/lastfm/api.js"></script>
-  <script src="/js/lastfm.obj.js"></script>
+  <script type="text/javascript" src="/js/commons.js"></script>
   <script src="/js/options.js"></script>
 </head>
 <body>
