+++ chrome-extension/popup.html
@@ -4,8 +4,7 @@
   <meta charset="utf-8">
   <link href="/css/generic.css" rel="stylesheet" type="text/css" />
   <link href="/css/popup.css" rel="stylesheet" type="text/css" />
-  <script src="/config.json"></script>
-  <script src="/js/settings.js"></script>
+  <script type="text/javascript" src="/js/commons.js"></script>
   <script type="text/javascript" src="/js/popup.js"></script>
 </head>
 <body>
