+++ javascript/modules/config.js
@@ -1,0 +1,9 @@
+define('ultra.config', function(require) { 'use strict';
+  return {
+    lastFM: {
+      authUri : 'http://www.last.fm/api/auth/?api_key=',
+          key : '0b26cafa64819d3b41788dc848ec0926',
+          sig : '660eb973c51b3003c4ca14fbef23af33'
+    }
+  }
+});
