+++ package.json
@@ -21,7 +21,7 @@
     "radio",
     "ultra"
   ],
-  "author": "Alex Mercer",
+  "author": "Alexey Bobyrev",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/sfate/UltraFM/issues"
